+++ Changelog.md
@@ -5,13 +5,18 @@ All notable changes to this project will be documented in this file since v2.0.0
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.1.3] - 2022-09-??
+## [3.1.3] - 2022-09-26
+
+### Added
+
+- Output final service parameters and mounted files after dumping a composition
 
 ### Fixed
 
 - Allow empty optional services list.
 - Allow passing empty option value via universal options like `--with-web-root=""`.
 - Make upgrade.sh more compatible with the latest Git versions.
+- Setting non-interactive mode for the `magento:setup` command.
 
 
 ## [3.1.2] - 2022-09-08